+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+
+- better error for invalid query selector strings
+
 ## [0.7.0] - 2018-07-06
 
 ### Changed